Moving files to upgrade package
