+++ internal/pkg/agent/application/upgrade/service_update_linux_test.go
@@ -4,7 +4,7 @@
 
 //go:build linux
 
-package cmd
+package upgrade
 
 import (
 	"os"
